chore: Adjust visibility for TextBlockMeasureCache.Clear
